requirements.txt
python-telegram-bot==20.0
mega.py==1.0.8